+++ demo/taro-vue3-app/src/app.config.ts
@@ -1,7 +1,5 @@
 export default defineAppConfig({
-  pages: [
-    'pages/index/index'
-  ],
+  pages: ['pages/index/index', 'pages/index/test'],
   window: {
     backgroundTextStyle: 'light',
     navigationBarBackgroundColor: '#fff',