+++ rollup.config.js
@@ -1,6 +1,7 @@
 import typescript from '@rollup/plugin-typescript'
 import { nodeResolve } from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
+import json from '@rollup/plugin-json'
 // import { terser } from 'rollup-plugin-terser'
 import pkg from './package.json'
 // const isProd = process.env.NODE_ENV === 'production'
@@ -9,13 +10,14 @@ const isDev = process.env.NODE_ENV === 'development'
 /** @type {import('rollup').RollupOptions } */
 const sharedConfig = {
   plugins: [
+    json(),
     nodeResolve({
       preferBuiltins: true
     }),
     commonjs(),
     typescript({ tsconfig: './tsconfig.build.json', sourceMap: isDev, declaration: false })
   ],
-  external: [...(pkg.dependencies ? Object.keys(pkg.dependencies) : [])]
+  external: [...(pkg.dependencies ? Object.keys(pkg.dependencies) : []), 'webpack', 'loader-utils']
 }
 
 /** @type {Array<import('rollup').RollupOptions> } */
@@ -26,8 +28,8 @@ const config = [
       {
         file: pkg.main,
         format: 'cjs',
-        sourcemap: isDev
-        // exports: 'auto'
+        sourcemap: isDev,
+        exports: 'auto'
       }
       // { format: 'esm', file: pkg.module, sourcemap: isDev }
     ],
@@ -45,7 +47,33 @@ const config = [
       // { format: 'esm', file: pkg.module, sourcemap: isDev }
     ],
     ...sharedConfig
+  },
+  {
+    input: 'src/loader/jsx-rename-loader.ts',
+    output: [
+      {
+        file: 'dist/jsx-rename-loader.js',
+        format: 'cjs',
+        sourcemap: isDev,
+        exports: 'auto'
+      }
+      // { format: 'esm', file: pkg.module, sourcemap: isDev }
+    ],
+    ...sharedConfig
   }
+  // {
+  //   input: 'src/loader/vue-template-rename-loader.ts',
+  //   output: [
+  //     {
+  //       file: 'dist/vue-template-rename-loader.js',
+  //       format: 'cjs',
+  //       sourcemap: isDev,
+  //       exports: 'auto'
+  //     }
+  //     // { format: 'esm', file: pkg.module, sourcemap: isDev }
+  //   ],
+  //   ...sharedConfig
+  // }
 ]
 
 export default config