+++ src/base/BaseJsxPlugin/v5.ts
@@ -4,7 +4,9 @@ import type { Compiler } from 'webpack'
 import { styleHandler, jsxHandler, pluginName, getOptions, createInjectPreflight } from '@/shared'
 import { createReplacer } from '@/jsx/replacer'
 import type { IBaseWebpackPlugin } from '@/interface'
-
+import { NormalModule } from 'webpack'
+import path from 'path'
+const NS = 'jsx-rename-loader'
 /**
  * @issue https://github.com/sonofmagic/weapp-tailwindcss-webpack-plugin/issues/2
  */
@@ -12,6 +14,7 @@ import type { IBaseWebpackPlugin } from '@/interface'
 export class BaseJsxWebpackPluginV5 implements IBaseWebpackPlugin {
   options: Required<UserDefinedOptions>
   appType: AppType
+  static NS = NS
   constructor (options: UserDefinedOptions = {}, appType: AppType) {
     this.options = getOptions(options)
     this.appType = appType
@@ -24,8 +27,26 @@ export class BaseJsxWebpackPluginV5 implements IBaseWebpackPlugin {
     const { ConcatSource } = compiler.webpack.sources
     // react
     const replacer = createReplacer()
+    const isReact = true
+    const rule = {
+      loader: path.resolve(__dirname, `${NS}.js`),
+      options: {
+        replacer
+      },
+      ident: null,
+      type: null
+    }
     onLoad()
     compiler.hooks.compilation.tap(pluginName, (compilation) => {
+      if (isReact) {
+        NormalModule.getCompilationHooks(compilation).loader.tap(pluginName, (loaderContext, module) => {
+          if (jsMatcher(module.resource)) {
+            // unshift
+            module.loaders.unshift(rule)
+          }
+        })
+      }
+
       compilation.hooks.processAssets.tap(
         {
           name: pluginName,
@@ -46,9 +67,9 @@ export class BaseJsxWebpackPluginV5 implements IBaseWebpackPlugin {
               const source = new ConcatSource(css)
               compilation.updateAsset(file, source)
               onUpdate(file)
-            } else if (jsMatcher(file)) {
+            } else if (!isReact && jsMatcher(file)) {
               const rawSource = originalSource.source().toString()
-              const { code } = jsxHandler(rawSource, replacer, file)
+              const { code } = jsxHandler(rawSource, replacer)
               const source = new ConcatSource(code)
               compilation.updateAsset(file, source)
               // const sourceMapFileName = `${file}.map`